requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
pytest==8.3.2
httpx==0.27.2
